feature: message controller 구현

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -53,7 +53,7 @@ public class BasicUserService implements UserService {
         String fileName = binaryContentRequest.getName();
         String contentType = binaryContentRequest.getContentType();
         byte[] bytes = binaryContentRequest.getBytes();
-        long size = bytes.length;
+        int size = bytes.length;
 
         BinaryContent binaryContent = new BinaryContent(fileName, size, contentType, bytes);
         binaryContentRepository.save(binaryContent);
@@ -114,7 +114,7 @@ public class BasicUserService implements UserService {
         String fileName = binaryContentRequest.getName();
         String contentType = binaryContentRequest.getContentType();
         byte[] bytes = binaryContentRequest.getBytes();
-        long length = bytes.length;
+        int length = bytes.length;
 
         BinaryContent binaryContent = new BinaryContent(fileName,
             length,
